+++ xtcp-net-3/pstcp/tcp_listen.cpp
@@ -55,7 +55,7 @@ void listen_statecb(void * ignore)
 	int state;
 	int error = -1;
 
-	state = (int)ignore;
+	state = (int)(long)ignore;
 	if (state == EV_RUNSTOP) {
 		fprintf(stderr, "listen_stop\n");
 		drop_event(&_event);
